requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
psycopg2-binary
python-dotenv
faker
python-multipart
